Add a gif file and a start with the Login and SignUp Viewcontrollers
A gif image is added to the startscreen.
The LoginViewcontroller is finished. I have to test it now on correctness with database.
The SignUpViewcontroller doesn't work yet fully. A user can write the e-mail, username, password and repeat password and the cancel button works. The user is only not able to login and perform the segue.