Try for a search function, display of information, FireBaseConnection and FavoriteLists
First of all, I tried to get my information per country stored in a TableViewController. I'm still not succeeded. The problem that I have is my data. Because it are not simple string, but it are sometimes string, sometimes integers, sometimes arrays and sometimes dictionaries. The amount of them change per country and the TableViewController has to change it rows per section dynamically, but I can't fulfill that. Next to that I want to make a search function but, because I don't have an array with strings of countrynames, will not succeed as well. The last thing that is a complete disaster is to make lists of countries and be able to add, save, change and delete them.